app/requirements.txt
faiss_cpu==1.7.4
numpy==1.22.4
Pillow==9.1.1
torch==1.13.1
torchvision==0.14.1
uvicorn==0.17.6
fastapi == 0.78.0
pydantic == 1.9.1
boto3 == 1.24.9
transformers
python-multipart
